+++ content-aside.php
@@ -2,12 +2,17 @@
 /**
  * The template for displaying posts in the Aside post format on index and archive pages
  * @since WP-Forge 5.5.1.7
- * @version 6.2.1.2
+ * @version 6.2.1.3
  */
 ?>
 	<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
 		<div class="entry-meta-header">
-			<a class="entry-format" href="<?php echo esc_url( get_post_format_link( 'aside' ) ); ?>" title="View all Aside Posts"><span class="genericon genericon-aside"></span> <?php echo get_post_format_string( 'aside' ); ?></a>			
+			<?php  if( get_theme_mod( 'wpforge_cat_display','yes' ) == 'yes') { ?>
+				<?php  if( get_theme_mod( 'wpforge_cat_position','top' ) == 'top') { ?>
+					<?php wpforge_entry_meta_categories(); ?>
+				<?php } // end if ?>
+			<?php } // end if ?>
+			<a class="entry-format" href="<?php echo esc_url( get_post_format_link( 'aside' ) ); ?>" title="View all Aside Posts"><span class="genericon genericon-aside"></span> <?php echo get_post_format_string( 'aside' ); ?></a>
 			<?php wpforge_entry_meta_header(); ?>
 			<?php if ( comments_open() ) : ?>
 				<span class="genericon genericon-comment"></span> <?php comments_popup_link( '<span class="leave-reply">' . __( 'Comment', 'wp-forge' ) . '</span>', __( '1 Comment', 'wp-forge' ), __( '% Comments', 'wp-forge' ) ); ?>
@@ -22,8 +27,10 @@
 		</div><!-- .aside -->
 		<footer class="entry-meta">
 			<div class="entry-meta-footer">
-				<?php if( get_theme_mod( 'wpforge_cat_display' ) == 'yes' || get_theme_mod( 'wpforge_cat_position' ) == 'bottom') { ?>
-					<?php wpforge_bottom_meta_categories(); ?>
+				<?php  if( get_theme_mod( 'wpforge_cat_display' ) == 'yes') { ?>
+					<?php  if( get_theme_mod( 'wpforge_cat_position' ) == 'bottom') { ?>
+						<?php wpforge_bottom_meta_categories(); ?>
+					<?php } // end if ?>
 				<?php } // end if ?>
 				<?php  if( get_theme_mod( 'wpforge_tag_display','yes' ) == 'yes') { ?>
 					<?php wpforge_entry_meta_footer(); ?>