+++ content-chat.php
@@ -2,11 +2,11 @@
 /**
  * The template for displaying posts in the Chat post format on index and archive pages.
  * @since WP-Forge 5.5.1.7
- * @version 6.2.1.2
+ * @version 6.2.1.3
  */
 ?>
 
-<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>	
+<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
 	<header class="entry-header">
 		<?php  if( get_theme_mod( 'wpforge_cat_display','yes' ) == 'yes') { ?>
 			<?php  if( get_theme_mod( 'wpforge_cat_position','top' ) == 'top') { ?>
@@ -21,13 +21,13 @@
 		</h1>
 		<?php endif; // is_single() ?>
 		<div class="entry-meta-header">
-			<a class="entry-format" href="<?php echo esc_url( get_post_format_link( 'chat' ) ); ?>" title="View all Chat Posts"><span class="genericon genericon-chat"></span> <?php echo get_post_format_string( 'chat' ); ?></a>			
+			<a class="entry-format" href="<?php echo esc_url( get_post_format_link( 'chat' ) ); ?>" title="View all Chat Posts"><span class="genericon genericon-chat"></span> <?php echo get_post_format_string( 'chat' ); ?></a>
 			<?php wpforge_entry_meta_header(); ?>
 			<?php if ( comments_open() ) : ?>
 				<span class="genericon genericon-comment"></span> <?php comments_popup_link( '<span class="leave-reply">' . __( 'Comment', 'wp-forge' ) . '</span>', __( '1 Comment', 'wp-forge' ), __( '% Comments', 'wp-forge' ) ); ?>
 			<?php endif; // comments_open() ?>
 			<?php edit_post_link( __( 'Edit', 'wp-forge' ), '<span class="edit-link"><span class="genericon genericon-edit"></span> ', '</span>' ); ?>
-		</div><!-- end .entry-meta-header -->		
+		</div><!-- end .entry-meta-header -->
 	</header><!-- .entry-header -->
 	<div class="entry-content-post">
 		<?php the_content(); ?>
@@ -35,8 +35,10 @@
 	</div><!-- .entry-content -->
 		<footer class="entry-meta">
 			<div class="entry-meta-footer">
-				<?php if( get_theme_mod( 'wpforge_cat_display' ) == 'yes' || get_theme_mod( 'wpforge_cat_position' ) == 'bottom') { ?>
-					<?php wpforge_bottom_meta_categories(); ?>
+				<?php  if( get_theme_mod( 'wpforge_cat_display' ) == 'yes') { ?>
+					<?php  if( get_theme_mod( 'wpforge_cat_position' ) == 'bottom') { ?>
+						<?php wpforge_bottom_meta_categories(); ?>
+					<?php } // end if ?>
 				<?php } // end if ?>
 				<?php  if( get_theme_mod( 'wpforge_tag_display','yes' ) == 'yes') { ?>
 					<?php wpforge_entry_meta_footer(); ?>