+++ inc/custom-header.php
@@ -2,7 +2,7 @@
 /**
  * Implement an optional custom header for WP-Forge
  * @since WP-Forge 5.5.1.7
- * @version 6.2.1.2
+ * @version 6.2.1.3
  */
 
 if ( ! function_exists( 'wpforge_custom_header_setup' ) ) {